Hack in AIS support
Hacky solution to support AIS at the expense of ADS-B. SDR devices can
be chosen for the purpose of AIS (and actually, they are always used for
that purpose due to a hack).

When at least one device is selected for AIS, AIS-catcher is started via
a compose file. Since AIS-catcher can't be configured via environment
variables, a small script changes its config file to use the correct SDR
serial.

When no devices are selected for ADS-B (1090, 978 MHz), the ultrafeeder
container is not started.