Show mDNS status in hotspot
On the hotspot page, show a status box that indicates whether mDNS is
likely to work once the feeder is connected to the network. To do this,
start the avahi service with the hotspot and try to reach the feeder as
adsb-feeder.local. Use a new /healthz endpoint that just returns a 200.
A little script on the hotspot page tries this for 5 minutes and assumes
mDNS isn't working if it's not successful by then.

We also have to modify the fake DNS server to never answer queries for
.local or .local.adsb-feeder.im.