Detect busybox more reliably
The install script used to check for limited functionality of busybox
commands by invoking them with the functionality that was needed. But
that could lead to wrong results when something unrelated wasn't working
right (e.g. USB for the lsusb -v check).

Call --help and see if it says BusyBox in there instead.