Disable nftables
The nftables firewall is causing problems with Docker and serving the
app. While it would be possible to define rules to access the app (and
DNS in hotspot mode), Docker doesn't natively support adding rules to
properly expose containers to the outside in nftables (only iptables).

Just turn it off whenever one of the services is running.