Shut down cleanly
Ever since introducing the systemd-inhibit wrapper, when stopping
adsb-setup, systemd would send a signal to the wrapper which never made
it to the actual app. Instead, it would get killed without notice.

To fix this, have the app write a PID file, and use an ExecStop script
to read it and shut down gracefully. As a backup, systemd will always
kill the whole process if the script is ineffective.