Regularly disable hotspot
While the hotspot is running, the wifi device is blocked. So if the
problem of missing connectivity was upstream (e.g. the router lost
connection), we have no way of seeing if it got fixed.

To address this, the manager now regularly switches from the hotspot
back to the main app for a while to see if connectivity returns.