Add install library
It would be nice to put functions to install and uninstall the app into
a common library, so we don't have to copy-paste around between the
install script and the updated functionality. But we also need a fully
self-contained install script that people can download.

Generate this script by replacing the sourcing of the library with its
contents, and do this in the CI pipeline.