Remove extra link variables
Some of the aggregators needed to pass extra information back to the API
to return as part of the response. This was done by writing it to the
config, only to be read again immediately. No other purpose.

The data is now passed as part of the status, and this hack is no longer
needed.