Use Aishub aggregator
Set the Aishub UDP port in the shipfeeder container to activate the
export.

Unfortunately, we need a bit of a hack in order to be able to turn
Porttracker and Aishub on and off independently of each other. Since
they're running from the same container, we have to set the
corresponding environment variables to empty strings to deactivate them.
But that means we lose our settings every time, which is not acceptable.
Instead, we introduce a second set of variables that are used for the
container, which are set or unset on every toggle of the status based on
the real variables that remain unchanged.

This is a workaround that we can probably find a better solution for
with the new config setup, where we can define settings that are only
written to the env file if a flag is set.