Add generated settings
We had to employ a hack in order to disable and enable AIS aggregators
independently from one another without deleting the configuration every
time. Since the aggregators are all running in the same container, they
must be turned off by setting their appropriate environment variables to
empty strings. Since we didn't want to throw away the configuration,
that means we had to maintain 2 sets of variables, one for storing the
information, and one to use in the compose file.

With generated settings, we can have just the one setting to store the
data (e.g. a sharing key), and one generated setting that automatically
copies that sharing key if the aggregator is enabled, or sets itself to
the empty string otherwise.