Fix checkbox parsing
Normally, unchecked checkboxes are simply left out of a submitted form.
But I'd overlooked a hook for form submit that's included in every page
via the base.html, which rewrites the values so that every checkbox is
included, with values exactly "0" or "1".

This is now handled properly and also documented in the README.