Refactor Ultrafeeder config generation
Instead of having this separate confusing class with essentially a
static method and writing a value to disk on every update, generate the
Ultrafeeder config on the fly.

This requires a bit of refactoring, as the System instance now has to be
started and the aggregators initialized before we access that generated
setting for the first time.