Write env file only if necessary
The env file used to get written a lot of times, e.g. when some setting
changed. This is an artifact from the old messy config setup. With the
new authoritative settings file, we only need to write the env file just
before it's actually read somewhere. At the moment, that's in the docker
compose wrapper script and in the Prometheus push service.