Refactor image pulling and pruning
There were several scripts and services involved in pulling necessary
images and pruning unused ones. Replaced this by a single service and
script that is WantedBy adsb-setup, so that this happens whenever the
devices starts, but also when an update is installed. If it should
become necessary in the future, this unit can also be started from the
application.