Handle multiple aggregator selection better
The setup page used to have a radio selection where you could say you
want to send to all aggregators which don't require an account, or that
you want to pick individually. That kind of makes sense, but it required
a bit of hackery in the back end where we had to fudge that flag around
in order to allow any changes to the aggregator selection.

Instead, replace this by 2 checkboxes where you can essentially enable
all aggregators with the submit if you want.