Fix mlat configuration
I have the suspicion that mlat/mlathub never worked the way as intended.
Using mlat and mlathub are different things. The MLAT_ENABLE environment
variable doesn't actually exist in ultrafeeder (anymore?), and it was
always set to enabled anyway.

Removed the unnecessary setting, and added a checkbox for mlathub to the
aggregators page.