Refactor implied settings
For the longest time, we've had the handle_implied_settings(), which was
called during every update and was basically fudging around in the
settings because they were not initialized correctly or an update broke
something.

With some refactoring and better use of the real settings, we can
finally get rid of this crutch.