Maintain regular mode with active gateway
This adds a new connectivity check that checks if there is at least one
gateway that responds to a ping. Effectively, this means that the feeder
will remain in regular mode (i.e. not start a hotspot) as long as it is
connected to a router, even if that router loses internet connection.