Improve graphical system detection
For the hotspot auto mode, we need to detect whether a graphical system
is available. This used to be done by checking if a system service was
running that looks like a window manager. But that failed because these
are enabled even on headless systems (as long as they're installed).

Instead, look at active processes and see if something that looks like X
or wayland sticks around for long enough.