Make SDR guesses per device
We have a bit of logic that guesses what a newly connected (and not yet
configured) SDR device could be used for. This used to assign each
purpose (ADS-B, UAT, AIS) exactly one device, which was limited when it
turned out that that purpose already has a device assigned.

Instead return for each device a list of possible uses in descending
order of preference.