Clear aggregator reminder
We have an info alert reminding the user after a fresh install to go to
the data sharing page and select aggregators to share with. But this
would only be cleared if the user chose to share with someone, otherwise
it would hang around.

Clear the alert in any case now, even if the user doesn't want to share
with anyone.