Add hash infix to static files
Up until now, we have the cachebust.sh script in the codebase, which
runs before every start, goes through all static files and renames them
to include their own hash. It then does a huge search and replace in all
html files to replace occurences of static file names to include the
hash. This all is done to ensure that, if a static file changes (mostly
during development), the new version is served immediately instead of
the old version from the browser cache (since the new version has a
different file name).

This is all a bit opaque, non-obvious, not to mention slow to run. To
achieve the same effect more explicitly, the Python application itself
now computes hashes for all static files on startup. A modified
url_for() function automatically modifies URLs to infix the file hash,
and a modified static view endpoint serves those files by checking and
then stripping out the hash first. The files as they exist on disk are
no longer modified.

As of this commit, the cachebust script still exists and is still
called, though.