Keep DHCP client running
We need to stop NetworkManager when we start up the hotspot. But that
includes our only DHCP client, so if anyone plugs in an ethernet cable
while the hotspot is up (which is all the time now), we can't get an IP
address. Add dhcpcd and start it whenever NetworkManager is down.